inheritance for libraries and partial inheritance for layouts, refactor theme generator to include clones, better handling of skins and other starterkits, make skins a real starterkit, add FileRename class, remove templates from starterkits make this an option in theme generator, ensure info required key type is always set, cleaned up some whitespace lines and probably some other things I forgot

+++ at_core/lib/Drupal/at_core/Helpers/FileStripReplace.php
@@ -17,9 +17,11 @@ class FileStripReplace {
    *   The replacement string.
    */
   public function fileStrReplace($file_path, $find, $replace) {
-    $file_contents = file_get_contents($file_path);
-    $file_contents = str_replace($find, $replace, $file_contents);
-    file_put_contents($file_path, $file_contents);
+    if (file_exists($file_path)) {
+      $file_contents = file_get_contents($file_path);
+      $file_contents = str_replace($find, $replace, $file_contents);
+      file_put_contents($file_path, $file_contents);
+    }
   }
 
 } // end class
